Move methods and hooks from App to TrailIndex
