+++ src/components/App/App.js
@@ -1,25 +1,15 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import NavBar from '../NavBar/NavBar';
 import TrailIndex from '../TrailIndex/TrailIndex';
 import TrailDetails from '../TrailDetails/TrailDetails';
-import Filter from '../Filter/Filter';
 import FavoriteTrails from '../FavoriteTrails/FavoriteTrails';
 import PageNotFound from '../PageNotFound/PageNotFound';
 import { Route, Switch } from 'react-router-dom';
-import trails from '../../sampleTrailData';
-import { filterByCatagories, cleanFilters } from '../Filter/helperMethods';
 import './App.scss';
 
 const userLoggedIn = {id: 23, name: 'Eric'}
 
 const App = () => {
-  const [allTrails, setAllTrails] = useState(trails);
-  const [filteredTrails, setFilteredTrails] = useState(trails);
-
-  const handleTrailFilters = (filterObj) => {
-    const cleanedFilters = cleanFilters(filterObj)
-    setFilteredTrails(filterByCatagories(cleanedFilters, allTrails))
-  }
 
   return (
     <main className="app-main">
@@ -27,22 +17,20 @@ const App = () => {
       <Switch>
         <Route exact path="/"
           render={() =>
-            <>
-              <input type="text" placeholder="Search Trails"></input>
-              <Filter handleTrailFilters={handleTrailFilters}/>
-              <TrailIndex filteredTrails={filteredTrails} />
+            <>  
+              <TrailIndex />
             </>
           }
         />
         <Route exact path="/favorites/:userID"
           render={({match}) =>
             <FavoriteTrails userID={match.params.userID}
-            trails={filteredTrails}/>
+            />
           }
         />
         <Route exact path="/trail/:id"
           render={({match}) => {
-              return <TrailDetails trailID={match.params.id} trails={filteredTrails}/>
+              return <TrailDetails trailID={match.params.id} />
           }}
         />
         <Route path="*"
